requirements.txt
spotipy
python-dotenv
boto3
pandas
fastapi
uvicorn
streamlit
